requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blessed==1.19.1
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
cachetools==5.3.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.26.4
comm==0.1.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.9.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gdown==4.6.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gpustat==1.0.0
grpcio==1.56.0
h5py==3.6.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.10.2
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.3.1
jupyter_server==2.2.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
librosa==0.8.1
lightning-fabric==2.0.4
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
musdb==0.4.0
museval==0.4.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy==1.22.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.495.46
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pooch==1.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyDeprecate==0.3.1
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.4
pytorch-lightning==1.5.9
pytz @ file:///croot/pytz_1671697431263/work
PyYAML==6.0
pyzmq==25.1.0
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
simplejson==3.18.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
stack-data==0.6.2
stempeg==0.2.3
sympy==1.12
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==1.13.1
torchaudio==0.13.1
torchcontrib==0.0.2
torchlibrosa==0.0.9
torchmetrics==0.11.1
torchvision==0.14.1
tornado==6.3.2
tqdm==4.62.3
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.5
yarl==1.8.2
zipp==3.12.0
